+++ NEWS.md
@@ -15,6 +15,7 @@
 * Add checkbox to databaseView to select all packages for reportDownload handler (#649)
 * Build Report Updates (#526)
 * Add "Explore Metrics" button to Uploaded Pkgs Summary tab on Upload Packages tab (#653)
+* Add comment module to the Function Explorer (#643)
 
 # riskassessment 2.0.0